Reorders total duration and braking RFD on the unilateral analysis pane. Also correctly renames unweighting RFD ==> braking RFD.
Adds the functionality to save analyzed data. Wherever the app file is saved (or wherever your project exists in the case of running the app via runGitHub()), a folder will be created called "Analyses" if it doesn't already exist. Depending on whether you're analyzing SJ or CMJ, 2 - 3 files will be created in this folder based on the date you select in the sidebar panel.